requirements_main_packages.txt
django
djangorestframework
drf-yasg
pytest-django
web3
eth-tester
py-evm
